1、新增Company的相关服务层、实现层、Dao层。 2、将account、user相关服务处、实现层、Dao层的数据转移到Company相应的位置。
