+++ ccbs-database/src/main/java/team/yingyingmonster/ccbs/database/mapper/zhw/ZhaoCompanyMapper.java
@@ -1,0 +1,10 @@
+package team.yingyingmonster.ccbs.database.mapper.zhw;
+
+import org.springframework.stereotype.Repository;
+import team.yingyingmonster.ccbs.database.bean.Company;
+
+@Repository
+public interface ZhaoCompanyMapper {
+    Integer addCompany(Company company);
+    Company selectCompany(Long accountId);
+}
