+++ ccbs-service/src/main/java/team/yingyingmonster/ccbs/service/serviceinterface/CompanyService.java
@@ -1,0 +1,7 @@
+package team.yingyingmonster.ccbs.service.serviceinterface;
+
+import team.yingyingmonster.ccbs.database.bean.Company;
+
+public interface CompanyService {
+    Company selectCompany(Long accountId);
+}
