+++ ccbs-service/src/main/java/team/yingyingmonster/ccbs/service/serviceinterface/UserService.java
@@ -1,0 +1,10 @@
+package team.yingyingmonster.ccbs.service.serviceinterface;
+
+import team.yingyingmonster.ccbs.database.bean.User;
+
+import java.util.List;
+
+public interface UserService {
+    Integer insertUsers(Long companyId, List<String[]> userList);//插入体检人员信息
+    List<User> selectUsers(Long companyId);//查询已插入体检人员名单
+}
